Routes extracted to dedicated files
