+++ frontend/lendme/lib/screens/home/items.dart
