+++ frontend/lendme/lib/screens/other/error.dart
