+++ frontend/lendme/lib/screens/other/splash.dart
